+++ CHANGES.rst
@@ -42,6 +42,9 @@ astropy.io.votable
 astropy.modeling
 ^^^^^^^^^^^^^^^^
 
+- Fix the shape of the outputs when a model set is evaluated with
+  ``model_set_axis=False`` . [#7317]
+
 astropy.nddata
 ^^^^^^^^^^^^^^